refactor(concepts): split concept-build engine out of endpoints/concepts.py
Extracts the synchronous concept-DB singleton, paged/targeted memory-row
reads, and the entity/relationship/code-link extraction loop (_run_build)
into services/concept_build_engine.py, continuing the v2.30.0 CLI/Console
module splits. Both the manual marm_concept_build route and the background
worker's build_for_memory_ids still drive this same engine via
asyncio.to_thread; endpoints/concepts.py keeps the FastAPI routes,
build-run bookkeeping, and async orchestration as the owner.

No behavior change. concepts.py drops from 789 to 446 lines. Retargets
tests that monkeypatch is_graph_available/extract_entities/_try_embed/
CONCEPT_BUILD_ROW_CAP/find_code_match/the concept-db singleton to the
module that now actually reads them at call time -- several of those
were previously patching endpoints.concepts, which no longer has any
effect on _run_build's real call sites.

Bumps to v2.39.1 (pure internal refactor, no user-facing change),
following the v2.21.2/v2.30.0 precedent for PATCH-only module splits.
