+++ docs/INSTALL-LINUX.md
@@ -313,7 +313,7 @@ curl -s http://localhost:8001/health
 {
   "status": "healthy",
   "service": "MARM MCP Server",
-  "version": "2.39.0",
+  "version": "2.39.1",
   "timestamp": "2026-01-01T00:00:00+00:00",
   "database": "connected",
   "semantic_search": "available"
