+++ marm-mcp-server/marm_mcp_server/__init__.py
@@ -14,12 +14,12 @@ Features:
 - Production-grade performance
 
 Author: Ryan Lyell - marm-memory
-Version: 2.39.0
+Version: 2.39.1
 """
 
 from typing import Any
 
-__version__ = "2.39.0"
+__version__ = "2.39.1"
 __author__ = "Ryan Lyell"
 __email__ = "ryanlyell@marmemory.com"
